Fix: Emit non-Ascii characters correctly
When the data from the webpage contains a non-Ascii character, LWP is
correctly decoding it from UTF-8. Our output stream also needs to be UTF-8
(or whatever the user is using) in order to print such characters
correctly.

This is for Issue #1, which it partially fixes.

Only partially, because accents are displayed wrongly on the webpage we're
scraping from! So this commit gets us from having unknown-character
placeholders to having accents following the character they're supposed to
be modifying, like on the website.

But it still doesn't actually combine the accents, as desired.