Refactoring: Web::Query instead of Web::Scraper
The creator of Web::Scraper recommends Web::Query:
https://twitter.com/miyagawa/status/1407052593097027590

I'm not sure that it's easier to use, but it's definitely more succinct,
because values are grabbed directly with queries. Whereas with
Web::Scraper, there's two stages: the scraping spec for storing values in
a data-structure, then taking them out again for processing.

Web::Query does away with that intermediate data structure, just
extracting elements and things from them as required.

I've also done away with some comments: when something is a single line,
it's often easier just to read it rather than have comments cluttering
everything up. You can tell what defines the title, because it's the line
setting $title based on a CSS selector.

Produces identical output to the previous version on the several accounts
I've tried it on.